requirements.txt
pymupdf>=1.23.0
Pillow>=10.0.0
numpy>=1.24.0
ttkbootstrap>=1.10.0
